add marshmallow expiriment for schema, pipeline
This commit adds a depedency on Marshmallow[1], a lightweight serialization/deserialization library that can also handle validation, requiring fields, etc. Additionally, I have made an attempt to set up simple pipeline building through method chaining. For example, building a pipeline might look like this:

my_pipeline = Pipeline().extractor(CsvExractor).schema(MySchema).load(Datapusher)

This will instantiate a new Pipeline, which would then be able to be either scheduled or run as expected. That functionality is to be added later. I've added a sample job taken from the Fatal OD in the existing WPRDC codebase for an example of what this would look like (and a skeleton test for it as well).

[1] http://marshmallow.readthedocs.org/en/latest/